src/main/java/com/oluwayanmi/library/controller/AuthorController.java
package com.oluwayanmi.library.controller;

import com.oluwayanmi.library.model.Author;
import com.oluwayanmi.library.service.AuthorService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import java.util.List;

@RestController
@RequestMapping("api/authors")
public class AuthorController {

    private final AuthorService authorService;

    public AuthorController(AuthorService authorService) {
        this.authorService = authorService;
    }

    @GetMapping("/{id}")
    public ResponseEntity<Author> getAuthorById(@PathVariable Long id) {
        Author author = authorService.getAuthorById(id);
        return ResponseEntity.ok(author);
    }

    @GetMapping
    public ResponseEntity<List<Author>> getAllAuthors() {
        List<Author> authors = authorService.getAllAuthors();
        return ResponseEntity.ok(authors);
    }

    @GetMapping("/{id}/books")
    public ResponseEntity<Author> getAuthorWithBooks(@PathVariable Long id) {
        Author author = authorService.getAuthorWithBooks(id);
        return ResponseEntity.ok(author);
    }

    @PostMapping(consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
    public ResponseEntity<Author> addAuthor(
            @RequestParam("name") String name,
            @RequestParam("bio") String bio,
            @RequestParam(value = "authorImage", required = false) MultipartFile authorImage) {

        Author author = new Author();
        author.setName(name);
        author.setBio(bio);

        Author savedAuthor = authorService.addAuthor(author, authorImage);
        return ResponseEntity.ok(savedAuthor);
    }

    @PutMapping("/{id}")
    public ResponseEntity<Author> updateAuthor(@PathVariable Long id, @RequestBody Author authorDetails) {
        Author updatedAuthor = authorService.updateAuthor(id, authorDetails);
        return ResponseEntity.ok(updatedAuthor);
    }

    @DeleteMapping("/{id}")
    public ResponseEntity<Void> deleteAuthor(@PathVariable Long id) {
        authorService.deleteAuthor(id);
        return ResponseEntity.noContent().build();
    }
}